Forgot some gint opacity values:
2002-03-04  Michael Natterer  <mitch@gimp.org>

	Forgot some gint opacity values:

	* app/core/gimplayer.[ch]: layer->opacity, gimp_layer_new(),
	gimp_layer_new_from_tiles()

	* app/core/gimpimage-projection.[ch]: gimp_image_projection_opacity()

	* app/core/gimpdrawable-transform.c
	* app/core/gimpedit.c
	* app/core/gimpimage-mask.c
	* app/core/gimpimage-merge.c
	* app/core/gimpimage-new.c
	* app/core/gimpimage.c
	* app/core/gimplayer-floating-sel.c
	* app/gui/layers-commands.c
	* app/tools/gimptexttool.c
	* app/widgets/gimplayerlistview.c
	* app/xcf/xcf-load.c
	* app/xcf/xcf-save.c
	* tools/pdbgen/pdb/layer.pdb: changed accordingly.

	* app/pdb/layer_cmds.c
	* libgimp/gimplayer_pdb.c: regenerated.

+++ app/core/gimpimage-projection.h
@@ -26,7 +26,7 @@ void            gimp_image_projection_free         (GimpImage       *gimage);
 TileManager   * gimp_image_projection              (GimpImage       *gimage);
 GimpImageType	gimp_image_projection_type         (const GimpImage *gimage);
 gint            gimp_image_projection_bytes        (const GimpImage *gimage);
-gint            gimp_image_projection_opacity      (const GimpImage *gimage);
+gdouble         gimp_image_projection_opacity      (const GimpImage *gimage);
 
 guchar        * gimp_image_projection_get_color_at (GimpImage       *gimage,
                                                     gint             x,
